Add cross-platform portability and code quality improvements
- Add MockRadarSensor for hardware-free testing (AMBIENT_MOCK_RADAR=true)
- Add cross-platform port auto-detection (Linux + macOS)
- Add .env.example with full configuration reference
- Fix all mypy type errors (12 -> 0)
- Fix all ruff lint warnings (5 -> 0)
- Add public properties to VitalsExtractor for filter access
- Update all debug scripts to use auto-detection
- Make Docker device paths configurable via environment
- Add README badges and improve documentation

+++ src/ambient/processing/fall_detection.py
@@ -132,7 +132,7 @@ class TrackHistory:
 		if not recent_heights:
 			return self.heights[-1], self.heights[-1], self.heights[-1]
 
-		return min(recent_heights), max(recent_heights), np.mean(recent_heights)
+		return min(recent_heights), max(recent_heights), float(np.mean(recent_heights))
 
 	def get_velocity_magnitude(self) -> float:
 		"""Get current velocity magnitude."""
